src/services/health/health-monitoring.service.ts
/**
 * @fileoverview TS 文件 health-monitoring.service.ts 的功能描述
 * @author Team
 * @copyright 2024 组织名称
 * @license ISC
 */

@Injectable()
export class HealthMonitoringService {
  // 生理指标监测
  vitalSigns: {
    // 基础指标
    basicMetrics: {
      bloodPressure: {
        systolic: number;
        diastolic: number;
      };
      heartRate: number;
      bodyTemperature: number;
      respiratoryRate: number;
      bloodOxygen: number;
    };

    // 代谢指标
    metabolicMetrics: {
      bloodSugar: number;
      cholesterol: {
        total: number;
        hdl: number;
        ldl: number;
      };
      bodyComposition: {
        weight: number;
        bmi: number;
        bodyFat: number;
        muscleMass: number;
      };
    };
  };

  // 健康评估
  healthAssessment: {
    // 风险评估
    riskAnalysis: {
      cardiovascular: number;
      diabetes: number;
      obesity: number;
      mentalHealth: number;
    };

    // 健康趋势
    trends: {
      shortTerm: HealthTrend[];
      longTerm: HealthTrend[];
      predictions: HealthPrediction[];
    };

    // 干预建议
    interventions: {
      lifestyle: string[];
      diet: string[];
      exercise: string[];
      medical: string[];
    };
  };
}
